+++ sqlalchemy_continuum/manager.py
@@ -1,6 +1,5 @@
 import re
 from functools import wraps
-from inflection import underscore, pluralize
 
 import sqlalchemy as sa
 from sqlalchemy.orm import object_session
@@ -86,7 +85,6 @@ class VersioningManager(object):
             'transaction_column_name': 'transaction_id',
             'end_transaction_column_name': 'end_transaction_id',
             'operation_type_column_name': 'operation_type',
-            'relation_naming_function': lambda a: pluralize(underscore(a)),
             'strategy': 'validity'
         }
         if plugins is None: