+++ sqlalchemy_continuum/__init__.py
@@ -16,7 +16,7 @@ from .utils import (
 )
 
 
-__version__ = '1.0'
+__version__ = '1.0.1'
 
 
 versioning_manager = VersioningManager()