+++ tests/test_delete.py
@@ -1,3 +1,4 @@
+import sqlalchemy as sa
 from tests import TestCase
 
 
@@ -23,3 +24,21 @@ class TestDelete(TestCase):
         assert len(versions) == 2
         assert versions[1].name == u'Some article'
         assert versions[1].content == u'Some content'
+
+
+class TestDeleteWithDeferredColumn(TestCase):
+    def create_models(self):
+        class TextItem(self.Model):
+            __tablename__ = 'text_item'
+            __versioned__ = {}
+            id = sa.Column(sa.Integer, autoincrement=True, primary_key=True)
+            name = sa.orm.deferred(sa.Column(sa.Unicode(255)))
+
+        self.TextItem = TextItem
+
+    def test_pop_with_delete(self):
+        item = self.TextItem()
+        self.session.add(item)
+        self.session.commit()
+        self.session.delete(item)
+        self.session.commit()